+++ packages/tendermint-rpc/src/rpcclients/httpbatchclient.ts
@@ -10,11 +10,16 @@ import { HttpEndpoint } from "./httpclient";
 import { hasProtocol, RpcClient } from "./rpcclient";
 
 export interface HttpBatchClientOptions {
+  /** Interval for dispatching batches (in milliseconds) */
   dispatchInterval: number;
+  /** Max number of items sent in one request */
   batchSizeLimit: number;
 }
 
-export const defaultHttpBatchClientOptions: HttpBatchClientOptions = {
+// Those values are private and can change any time.
+// Does a user need to know them? I don't think so. You either set
+// a custom value or leave the option field unset.
+const defaultHttpBatchClientOptions: HttpBatchClientOptions = {
   dispatchInterval: 20,
   batchSizeLimit: 20,
 };
@@ -31,11 +36,11 @@ export class HttpBatchClient implements RpcClient {
     reject: (a: Error) => void;
   }> = [];
 
-  public constructor(
-    endpoint: string | HttpEndpoint,
-    options: HttpBatchClientOptions = defaultHttpBatchClientOptions,
-  ) {
-    this.options = options;
+  public constructor(endpoint: string | HttpEndpoint, options: Partial<HttpBatchClientOptions> = {}) {
+    this.options = {
+      batchSizeLimit: options.batchSizeLimit ?? defaultHttpBatchClientOptions.batchSizeLimit,
+      dispatchInterval: options.dispatchInterval ?? defaultHttpBatchClientOptions.dispatchInterval,
+    };
     if (typeof endpoint === "string") {
       // accept host.name:port and assume http protocol
       this.url = hasProtocol(endpoint) ? endpoint : "http://" + endpoint;
@@ -55,6 +60,11 @@ export class HttpBatchClient implements RpcClient {
   public async execute(request: JsonRpcRequest): Promise<JsonRpcSuccessResponse> {
     return new Promise((resolve, reject) => {
       this.queue.push({ request, resolve, reject });
+
+      if (this.queue.length >= this.options.batchSizeLimit) {
+        // this train is full, let's go
+        this.tick();
+      }
     });
   }
 
@@ -68,27 +78,43 @@ export class HttpBatchClient implements RpcClient {
     }
   }
 
-  private async tick(): Promise<void> {
+  /**
+   * This is called in an interval where promise rejections cannot be handled.
+   * So this is not async and HTTP errors need to be handled by the queued promises.
+   */
+  private tick(): void {
     // Avoid race conditions
-    const queue = this.queue.splice(0, this.options.batchSizeLimit);
+    const batch = this.queue.splice(0, this.options.batchSizeLimit);
 
-    if (!queue.length) return;
+    if (!batch.length) return;
 
-    const request = queue.map((s) => s.request);
-    const raw = await http("POST", this.url, this.headers, request);
-    // Requests with a single entry return as an object
-    const arr = Array.isArray(raw) ? raw : [raw];
+    const requests = batch.map((s) => s.request);
+    const requestIds = requests.map((request) => request.id);
 
-    arr.forEach((el) => {
-      const req = queue.find((s) => s.request.id === el.id);
-      if (!req) return;
-      const { reject, resolve } = req;
-      const response = parseJsonRpcResponse(el);
-      if (isJsonRpcErrorResponse(response)) {
-        reject(new Error(JSON.stringify(response.error)));
-      } else {
-        resolve(response);
-      }
-    });
+    http("POST", this.url, this.headers, requests).then(
+      (raw) => {
+        // Requests with a single entry return as an object
+        const arr = Array.isArray(raw) ? raw : [raw];
+
+        arr.forEach((el) => {
+          const req = batch.find((s) => s.request.id === el.id);
+          if (!req) return;
+          const { reject, resolve } = req;
+          const response = parseJsonRpcResponse(el);
+          if (isJsonRpcErrorResponse(response)) {
+            reject(new Error(JSON.stringify(response.error)));
+          } else {
+            resolve(response);
+          }
+        });
+      },
+      (error) => {
+        for (const requestId of requestIds) {
+          const req = batch.find((s) => s.request.id === requestId);
+          if (!req) return;
+          req.reject(error);
+        }
+      },
+    );
   }
 }